Reorganize <rox/server.h> into a bunch of header files.
That would be useful for building client SDK.
