+++ src/roxx/extensions.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "rox/api.h"
+#include "rox/macros.h"
 #include "parser.h"
 #include "core/repositories.h"
 #include "core/properties.h"
